DONE(Rotas & Controladores): Separados rotas e controladores de usuários e de publicações. Nesse padrão será mantido.
