+++ server/handles.user.js
@@ -1,28 +1,4 @@
-const publicacoes = require('./publicacoes')
 const usuarios = require('./usuarios')
-
-async function listarPublicacoes(req, res) {
-  // TODO: Adicionar limitação de publicações carregadas pelo usuário
-  return await publicacoes.listarPublicacoes()
-    .then((posts) => {
-      return res.response(posts).code(200)
-    })
-    .catch((err) => {
-      return res.response("Erro ao resgatar as publicações").code(401)
-    })
-
-}
-
-async function criarPublicacao(req, res) {
-  console.log(req.pay)
-  return await publicacoes.criarPublicacao(req.payload.txtNome, req.payload.txtTitulo, req.payload.txtDepoimento)
-    .then((post) => {
-      return res.response(post).code(201)
-    })
-    .catch((err) => {
-      res.response("Erro ao criar a publicação").code(401)
-    })
-}
 
 async function logarUsuario(req, res) {
   return await usuarios.logarUsuarios(req.payload.email, req.payload.senha)
@@ -30,6 +6,7 @@ async function logarUsuario(req, res) {
       return res.response(user).code(200)
     })
     .catch((err) => {
+      // TODO: Gerenciar melhor o erro e passar claramente o erro lançado
       const errou = "Erro ao logar os usuário!"
       console.warn(errou, err)
       return res.response(JSON.stringify({ err })).code(401)
@@ -50,8 +27,6 @@ async function criarUsuario(req, res) {
 }
 
 module.exports = {
-  listarPublicacoes,
-  criarPublicacao,
   logarUsuario,
   criarUsuario,
 }
