+++ server/routes.index.js
@@ -1,0 +1,3 @@
+const routesUser = require('./routes.user')
+const routesPost = require('./routes.post')
+module.exports = [...routesUser, ...routesPost]
