+++ server/routes.post.js
@@ -1,0 +1,20 @@
+const handlesPost = require('./handles.post')
+const isValidPost = require('./validacao')
+
+module.exports = [
+  {
+    method: 'GET',
+    path: '/api/v1/publicacoes/',
+    handler: handlesPost.listarPublicacoes
+  },
+  {
+    method: 'POST',
+    path: '/api/v1/publicacao/',
+    handler: handlesPost.criarPublicacao,
+    options: {
+      validate: {
+        payload: isValidPost.criarPost,
+      },
+    },
+  },
+]
